requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-telegram-bot==21.3
google-generativeai==0.7.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-api-python-client==2.133.0
